упростил, отрефакторил и некоторые детали улучшил
