+++ cs/TagsCloudVisualization/TagCloud/CloundBitMap.cs
@@ -12,14 +12,16 @@ public class CloundBitMap : ITagCloudImage
     public CloundBitMap(int width, int height, string filePath)
     {
         if (!Directory.Exists(Path.GetDirectoryName(filePath))) throw new DirectoryNotFoundException();
+        if (width <= 0 || height <= 0) throw new ArgumentException("size of image should be with positive number");
 
         bitmap = new Bitmap(width, height);
         graphics = Graphics.FromImage(bitmap);
         graphics.Clear(Color.Black);
         this.filePath = filePath;
     }
-    
+
     public Size Size() => bitmap.Size;
+
 
     public void Draw(Rectangle rec)
     {
